refactor: Restructure DI modules and move mappers
