--- composeApp/src/androidMain/kotlin/com/moneyplusplus/money/di/PlatformModule.android.kt
@@ -1,11 +1,0 @@
-package com.moneyplusplus.money.di
-
-import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.engine.android.Android
-import org.koin.core.module.Module
-import org.koin.dsl.module
-
-actual val platformModule: Module =
-    module {
-        single<HttpClientEngine> { Android.create() }
-    }
