+++ composeApp/src/commonMain/kotlin/com/moneyplusplus/money/di/InitKoin.kt
@@ -1,6 +1,7 @@
 package com.moneyplusplus.money.di
 
 import com.moneyplusplus.data.di.moneyDataModule
+import com.moneyplusplus.domain.di.domainModule
 import com.moneyplusplus.presentation.di.moneyPresentationModule
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration
@@ -19,7 +20,8 @@ fun doInitKoin() {
 
 val appModule = module {
     includes(
-        moneyDataModule
+        moneyDataModule,
         moneyPresentationModule,
+        domainModule
     )
 }
