--- composeApp/src/commonMain/kotlin/com/moneyplusplus/money/di/PlatformModule.kt
@@ -1,6 +1,0 @@
-package com.moneyplusplus.money.di
-
-import org.koin.core.module.Module
-
-
-expect val platformModule: Module
