+++ data/build.gradle.kts
@@ -21,6 +21,7 @@ kotlin {
             implementation(libs.ktor.client.content.negotiation)
             implementation(libs.ktor.client.logging)
             implementation(libs.koin.core)
+            implementation(libs.ktor.ktor.client.android.v340)
             implementation(libs.kotlinx.datetime)
         }
         commonTest.dependencies {
