+++ data/src/androidMain/kotlin/com/moneyplusplus/data/di/platformModule.android.kt
@@ -1,7 +1,8 @@
 package com.moneyplusplus.data.di
 
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.android.Android
 import org.koin.dsl.module
 
 internal actual fun platformModule() = module {
-    // Android-specific dependencies can be declared here
-}
+    single<HttpClientEngine> { Android.create() }}
