+++ data/src/iosMain/kotlin/com/moneyplusplus/data/di/platformModule.ios.kt
@@ -1,7 +1,9 @@
 package com.moneyplusplus.data.di
 
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.darwin.Darwin
 import org.koin.dsl.module
 
 internal actual fun platformModule() = module {
-    // iOS-specific dependencies can be declared here
+    single<HttpClientEngine> { Darwin.create() }
 }
