+++ presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/di/ViewModelModule.kt
@@ -1,9 +1,8 @@
 package com.moneyplusplus.presentation.di
 
 import com.moneyplusplus.presentation.feature.forgetPassword.ForgetPasswordViewModel
-import com.moneyplusplus.presentation.login.LoginViewModel
 import com.moneyplusplus.presentation.feature.home.HomeViewModel
-import com.moneyplusplus.presentation.feature.transaction.TransactionViewModel
+import com.moneyplusplus.presentation.login.LoginViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -11,5 +10,4 @@ val viewModelModule = module {
     viewModelOf(::ForgetPasswordViewModel)
     viewModelOf(::LoginViewModel)
     viewModelOf(::HomeViewModel)
-    viewModelOf(::TransactionViewModel)
 }
